# documentation/SwaggerOpenAPI3.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: GITFIT API
  description: The GitFit API
servers:
  - url: https://git-fit-cop4331.herokuapp.com/api
...
paths:
  /users/emailVerification/sendVerificationEmail:
    post:
      tags:
        - Account Stuff
      summary: Verify email address
      operationId: sendVerificationEmail
      description: Send verification email or something
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sendVerificationEmailRequest'
        description: eh, something, or something
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sendVerificationEmailResponse'
        '404':
          description: URL Not Found
        '500':
          description: Server Error
  /users/emailVerification/verifyEmail:
    post:
      tags:
        - Account Stuff
      summary: Verify email address
      operationId: verifyEmail
      description: Verify email or something
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/verifyEmailRequest'
        description: eh, something
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/verifyEmailResponse'
        '404':
          description: URL Not Found
        '500':
          description: Server Error
...
components:
  schemas:
    sendVerificationEmailRequest:
      type: object
      properties:
        userId:
          type: string
          example: 6248b3843e1b4c6ea572d56b
        email:
          type: string
          example: ke131012@gmail.com
    sendVerificationEmailResponse:
      type: object
      properties:
        error:
          type: integer
    verifyEmailRequest:
      type: object
      properties:
        userId:
          type: string
          example: 6248b3843e1b4c6ea572d56b
    verifyEmailResponse:
      type: object
      properties:
        error:
          type: integer
...